Now it's complete

+++ HashProject/HashProject/HashProject.cpp
@@ -22,13 +22,21 @@ void inserir() {
 	string name;
 
 	cout << "Digite o ra que deseja inserir: ";
-	cin >> ra;
+	cin >> ra;	
 	cout << endl;
-	cout << "Digite o nome que deseja inserir: ";
-	cin >> name;
-	cout << "\n";
 
-	tabela.AddItem(ra, name);
+	if (tabela.Exist(ra))
+		cout << "Esse elemento já existe!\n" << endl;
+	else {
+		cout << "Digite o nome que deseja inserir: ";
+		cin >> name;
+		cout << "\n";
+
+		tabela.AddItem(ra, name);
+
+		if (tabela.Exist(ra))
+			cout << "Item incluído com sucesso!\n" << endl;
+	}
 }
 
 void excluir() {
@@ -36,7 +44,15 @@ void excluir() {
 	cout << "Digite o RA que deseja excluir: ";
 	cin >> ra;
 	cout << "\n";
-	tabela.DeleteItem(ra);
+
+	if (!tabela.Exist(ra))
+		cout << "Esse item não existe!\n" << endl;
+	else {
+		tabela.DeleteItem(ra);
+		
+		if (!tabela.Exist(ra))
+			cout << "Item excluído com sucesso!\n" << endl;
+	}
 }
 
 int main()
